@@ -12,18 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.11",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
-    "@angular/fire": "^6.1",
+    "@angular/animations": "^14.2.4",
+    "@angular/common": "^14.2.4",
+    "@angular/compiler": "^14.2.4",
+    "@angular/core": "^14.2.4",
+    "@angular/forms": "^14.2.4",
+    "@angular/platform-browser": "^14.2.4",
+    "@angular/platform-browser-dynamic": "^14.2.4",
+    "@angular/router": "^14.2.4",
+    "@angular/fire": "^7.4",
     "bootstrap": "^4.3.1",
     "core-js": "^3.3.2",
-    "firebase": "^8.0",
+    "firebase": "^9.0",
     "modernizr": "^3.7.1",
     "ngx-siema": "^2.0.1",
     "ngx-toastr": "13.2.1",
@@ -33,10 +33,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/language-service": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.4",
+    "@angular/cli": "^14.2.4",
+    "@angular/compiler-cli": "^14.2.4",
+    "@angular/language-service": "^14.2.4",
     "@types/dotenv": "^6.1.1",
     "@types/jasmine": "^3.4.4",
     "@types/jasminewd2": "^2.0.8",
